Move all `embassy` examples to `async` and convert to individual projects
