+++ examples/async/embassy_hello_world/src/main.rs
@@ -3,9 +3,6 @@
 //! This is an example of running the embassy executor with multiple tasks
 //! concurrently.
 
-//% CHIPS: esp32 esp32c2 esp32c3 esp32c6 esp32h2 esp32s2 esp32s3
-//% FEATURES: embassy esp-hal/unstable
-
 #![no_std]
 #![no_main]
