+++ examples/async/embassy_multicore_interrupt/src/main.rs
@@ -1,4 +1,4 @@
-//! This example shows how to spawn async tasks on the second core.
+//! This example shows how to use the interrupt executors on either core.
 //!
 //! The second core runs a simple LED blinking task, that is controlled by a
 //! signal set by the task running on the other core.
@@ -6,24 +6,22 @@
 //! The following wiring is assumed:
 //! - LED => GPIO0
 
-//% CHIPS: esp32 esp32s3
-//% FEATURES: embassy esp-hal/unstable
-
 #![no_std]
 #![no_main]
 
 use core::ptr::addr_of_mut;
 
-use embassy_executor::Spawner;
 use embassy_sync::{blocking_mutex::raw::CriticalSectionRawMutex, signal::Signal};
 use embassy_time::{Duration, Ticker};
 use esp_backtrace as _;
 use esp_hal::{
     gpio::{Level, Output, OutputConfig},
+    interrupt::{Priority, software::SoftwareInterruptControl},
+    main,
     system::{Cpu, CpuControl, Stack},
     timer::{AnyTimer, timg::TimerGroup},
 };
-use esp_hal_embassy::Executor;
+use esp_hal_embassy::InterruptExecutor;
 use esp_println::println;
 use static_cell::StaticCell;
 
@@ -50,9 +48,31 @@ async fn control_led(
     }
 }
 
-#[esp_hal_embassy::main]
-async fn main(_spawner: Spawner) {
+/// Sends periodic messages to control_led, enabling or disabling it.
+#[embassy_executor::task]
+async fn enable_disable_led(control: &'static Signal<CriticalSectionRawMutex, bool>) {
+    println!(
+        "Starting enable_disable_led() on core {}",
+        Cpu::current() as usize
+    );
+    let mut ticker = Ticker::every(Duration::from_secs(1));
+    loop {
+        esp_println::println!("Sending LED on");
+        control.signal(true);
+        ticker.next().await;
+
+        esp_println::println!("Sending LED off");
+        control.signal(false);
+        ticker.next().await;
+    }
+}
+
+#[main]
+fn main() -> ! {
+    esp_println::logger::init_logger_from_env();
     let peripherals = esp_hal::init(esp_hal::Config::default());
+
+    let sw_ints = SoftwareInterruptControl::new(peripherals.SW_INTERRUPT);
 
     let timg0 = TimerGroup::new(peripherals.TIMG0);
     let timer0: AnyTimer = timg0.timer0.into();
@@ -66,29 +86,28 @@ async fn main(_spawner: Spawner) {
 
     let led = Output::new(peripherals.GPIO0, Level::Low, OutputConfig::default());
 
+    static EXECUTOR_CORE_1: StaticCell<InterruptExecutor<1>> = StaticCell::new();
+    let executor_core1 = InterruptExecutor::new(sw_ints.software_interrupt1);
+    let executor_core1 = EXECUTOR_CORE_1.init(executor_core1);
+
     let _guard = cpu_control
         .start_app_core(unsafe { &mut *addr_of_mut!(APP_CORE_STACK) }, move || {
-            static EXECUTOR: StaticCell<Executor> = StaticCell::new();
-            let executor = EXECUTOR.init(Executor::new());
-            executor.run(|spawner| {
-                spawner.spawn(control_led(led, led_ctrl_signal)).ok();
-            });
+            let spawner = executor_core1.start(Priority::Priority1);
+
+            spawner.spawn(control_led(led, led_ctrl_signal)).ok();
+
+            // Just loop to show that the main thread does not need to poll the executor.
+            loop {}
         })
         .unwrap();
 
-    // Sends periodic messages to control_led, enabling or disabling it.
-    println!(
-        "Starting enable_disable_led() on core {}",
-        Cpu::current() as usize
-    );
-    let mut ticker = Ticker::every(Duration::from_secs(1));
-    loop {
-        esp_println::println!("Sending LED on");
-        led_ctrl_signal.signal(true);
-        ticker.next().await;
+    static EXECUTOR_CORE_0: StaticCell<InterruptExecutor<0>> = StaticCell::new();
+    let executor_core0 = InterruptExecutor::new(sw_ints.software_interrupt0);
+    let executor_core0 = EXECUTOR_CORE_0.init(executor_core0);
 
-        esp_println::println!("Sending LED off");
-        led_ctrl_signal.signal(false);
-        ticker.next().await;
-    }
+    let spawner = executor_core0.start(Priority::Priority1);
+    spawner.spawn(enable_disable_led(led_ctrl_signal)).ok();
+
+    // Just loop to show that the main thread does not need to poll the executor.
+    loop {}
 }
