+++ examples/async/embassy_multiprio/src/main.rs
@@ -13,9 +13,6 @@
 // `low_prio_async` and `low_prio_blocking`. The interrupt-executor is created in `main` and is used
 // to spawn `high_prio`.
 
-//% CHIPS: esp32 esp32c2 esp32c3 esp32c6 esp32h2 esp32s2 esp32s3
-//% FEATURES: embassy esp-hal-embassy/log-04 esp-hal/unstable
-
 #![no_std]
 #![no_main]
