+++ examples/async/embassy_rmt_tx/src/main.rs
@@ -5,9 +5,6 @@
 //! The following wiring is assumed:
 //! - generated pulses => GPIO4
 
-//% CHIPS: esp32 esp32c3 esp32c6 esp32h2 esp32s2 esp32s3
-//% FEATURES: embassy esp-hal/unstable
-
 #![no_std]
 #![no_main]
 
@@ -27,6 +24,8 @@ esp_bootloader_esp_idf::esp_app_desc!();
 #[esp_hal_embassy::main]
 async fn main(_spawner: Spawner) {
     println!("Init!");
+
+    esp_println::logger::init_logger_from_env();
     let peripherals = esp_hal::init(esp_hal::Config::default());
 
     let timg0 = TimerGroup::new(peripherals.TIMG0);
